Add get, create, and remove pokemon on a team
Add the add_pokemon action on the PokemonTeamViewSet in order
to allow the creation of a new pokemon and add it to the team.

The manage_pokemon action allow users to sent a DELETE request
to remove a pokemon from the team. It does not delete the pokemon
from the database.

To allow the creation of nested objects the update and create
methods of PokemonSerializer were overwritten. Therefore
includding a list of abilities with the pokemon data will
create and add each ability to the pokemon instance.